Move usb utility functions to usb-util module
